코드 리팩토링 작업중

+++ app/src/main/java/org/jeonfeel/moeuibit2/ui/coindetail/newScreen/order/NewOrderScreen.kt
@@ -76,7 +76,7 @@ fun NewOrderScreen(
     changeOrderBookIndicationState: () -> Unit,
     getUserSeedMoney: () -> Long,
     getUserBTC: () -> Double,
-    requestBid: (String, Double, BigDecimal, Long) -> Unit,
+    requestBid: (String, Double, BigDecimal, Double) -> Unit,
     requestAsk: (String, Double, Long, BigDecimal) -> Unit,
     getUserCoin: () -> MyCoin,
     btcPrice: State<BigDecimal>
