+++ app/src/main/java/org/jeonfeel/moeuibit2/ui/coindetail/newScreen/order/OrderScreenRoute.kt
@@ -21,7 +21,7 @@ fun OrderScreenRoute(
     changeOrderBookIndicationState: () -> Unit,
     getUserSeedMoney: () -> Long,
     getUserBTC: () -> Double,
-    requestBid: (String, Double, BigDecimal, Long) -> Unit,
+    requestBid: (String, Double, BigDecimal, Double) -> Unit,
     requestAsk: (String, Double, Long, BigDecimal) -> Unit,
     getUserCoin: () -> MyCoin,
     btcPrice: State<BigDecimal>
